requirements.txt
paho-mqtt==1.5.0
rhasspy-hermes==0.1.5
requests
